feat: update bindings

+++ packages/taiko-client-rs/crates/proposer/src/transaction_builder.rs
@@ -89,7 +89,7 @@ impl ShastaProposalTransactionBuilder {
                 offset: U24::ZERO,
             },
             // Include all forced inclusions in the source manifest.
-            numForcedInclusions: u8::MAX,
+            numForcedInclusions: u16::MAX,
         };
 
         // Build the transaction request with blob sidecar.
